trace_review/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
langfuse
python-dotenv==1.0.0

# Authentication
PyJWT==2.8.0
requests==2.31.0
